Remove the `Alert` CRD
Currently the auto-healing mechanism uses the `Alert` object to decouple
the generator of the alert and the healer. This isn't considered a good
architecture, because it requires the intervention of the Kubernetes
control plane during the healing process. To address that issue this
patch remove sthe `Alert` object and modifies healer so that it will
receive the alerts directly from the Prometheus alert manager, using an
HTTP web hook.
