--- images/alert-publisher/Dockerfile.in
@@ -1,40 +1,0 @@
-#
-# Copyright (c) 2018 Red Hat, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-FROM centos:7
-
-# Create the monitoring user. The unusually high UID is needed because
-# recent versions of Kubernetes don't allow lower UID numbers.
-RUN \
-useradd -l -u 1000000000 -M monitoring
-
-# Install the alert publisher binary:
-COPY alert-publisher /tmp
-RUN \
-cd tmp && \
-install -o root -g root -m u+rwx,g+rx,o+rx alert-publisher /usr/bin && \
-rm -rf /tmp/*
-
-# Install the alert publisher configuration:
-RUN \
-cd tmp && \
-install -o monitoring -g monitoring -m u+rwx,g+rx,o+rx -d /var/lib/monitoring && \
-rm -rf /tmp/*
-
-EXPOSE 9099
-USER monitoring
-WORKDIR /var/lib/monitoring
-CMD ["/usr/bin/alert-publisher", "-alsologtostderr"]
