--- images/alert-publisher/Downloads.in
@@ -1,1 +1,0 @@
-f181f619c9a8e0750c1ac940eb00a0881cc50386d896f06f159e9a5b68db60a0 https://github.com/prometheus/prometheus/releases/download/v2.1.0/prometheus-2.1.0.linux-amd64.tar.gz prometheus.tar.gz
