--- manifests/example-alert.yml
@@ -1,11 +1,0 @@
-apiVersion: monitoring.openshift.io/v1alpha1
-kind: Alert
-metadata:
-  name: node-down
-status:
-  labels:
-    alertname: NodeDown
-    job: nodes
-    intance: node1
-  annotations:
-    message: "Node 'node1.openshift.jhernand.private' is down"
