+++ pkg/apis/monitoring/v1alpha1/register.go
@@ -55,8 +55,6 @@ var (
 //
 func addKnownTypes(scheme *runtime.Scheme) error {
 	scheme.AddKnownTypes(SchemeGroupVersion,
-		&Alert{},
-		&AlertList{},
 		&AlertingRule{},
 		&AlertingRuleList{},
 		&HealingRule{},
