+++ pkg/apis/monitoring/v1alpha1/types.go
@@ -25,81 +25,6 @@ import (
 // +genclient
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 
-// Alert is the description of an alert.
-//
-type Alert struct {
-	meta.TypeMeta `json:",inline"`
-
-	// Standard object metadata.
-	// +optional
-	meta.ObjectMeta `json:"metadata,omitempty" protobuf:"bytes,1,opt,name=metadata"`
-
-	// The status of the alert.
-	// +optional
-	Status AlertStatus `json:"status,omitempty" protobuf:"bytes,2,opt,name=status"`
-}
-
-// AlertStatus is the status for an alert.
-//
-type AlertStatus struct {
-	// Prometheus labels.
-	// +optional
-	Labels map[string]string `json:"labels,omitempty" protobuf:"bytes,1,rep,name=labels"`
-
-	// Prometheus annotations.
-	// +optional
-	Annotations map[string]string `json:"annotations,omitempty" protobuf:"bytes,2,rep,name=annotations"`
-
-	// Current state of the alert.
-	// +optional
-	Conditions []AlertCondition `json:"conditions,omitempty" protobuf:"bytes,3,rep,name=conditions"`
-}
-
-// AlertConditionType are the valid types of alert conditions.
-//
-type AlertConditionType string
-
-// Valid types of alert conditions.
-//
-const (
-	// The alert is firing.
-	AlertFiring AlertConditionType = "Firing"
-
-	// The alert is resolved.
-	AlertResolved AlertConditionType = "Resolved"
-
-	// The alert is being healed.
-	AlertHealing AlertConditionType = "Healing"
-)
-
-// AlertCondition contains details for the current condition of the alert.
-//
-type AlertCondition struct {
-	// The type of condition.
-	Type AlertConditionType `json:"type,omitempty" protobuf:"bytes,1,opt,name=type"`
-
-	// Human readable message containing details of the condition.
-	Message string `json:"message,omitempty" protobuf:"bytes,2,opt,name=message"`
-}
-
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// AlertList is a list of alerts.
-//
-type AlertList struct {
-	meta.TypeMeta `json:",inline"`
-
-	// Standard list metadata.
-	// +optional
-	meta.ListMeta `json:"metadata,omitempty" protobuf:"bytes,1,opt,name=metadata"`
-
-	// List of alerts.
-	Items []Alert `json:"items" protobuf:"bytes,2,rep,name=items"`
-}
-
-// +genclient
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
 // AlertRule is the description of an alerting rule.
 //
 type AlertingRule struct {
