+++ playbooks/ensure_ovirt_node.yml
@@ -14,27 +14,14 @@
 # limitations under the License.
 #
 
-# This custom type definition is used to represent a firing alert.
+---
+- name: Ensure that an oVirt host is running
+  hosts: localhost
+  connection: local
+  gather_facts: false
 
-apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  name: alerts.monitoring.openshift.io
-spec:
-  group: monitoring.openshift.io
-  version: v1alpha1
-  names:
-    kind: Alert
-    listKind: AlertList
-    singular: alert
-    plural: alerts
-  scope: Namespaced
-  validation:
-    openAPIV3Schema:
-      properties:
-        status:
-          properties:
-            annotations:
-              type: object
-            labels:
-              type: object
+  tasks:
+  - name: Ensure oVirt host {{ ovirt_host }}
+    ovirt_vms:
+      name: jhernand-openshift-{{ ovirt_host | regex_replace('\..*$', '') }}
+      state: running
